Add raise_if_exists flag to hdfs_client.mkdir()
It is *way* to complicated to actually implement it. So I just went with
assertions to make sure that the programmers who want can explicitly
state their intent.  See spotify/luigi#393 for a discussion about this.

Oh, and I also started to use it at one place to get the ball rolling.


Former-commit-id: c6f144d1696a3c95c44e949e7d9c32ad6b0c03d2