Add support for adding egg-info files to packages archive.
In order for packages that make use of entry points to work on Hadoop
node, the relevant egg-info directories also need to be copied to those
nodes. This change performs a simple check to see if a related egg-info
is present for a requested package, and adds those files as well.


Former-commit-id: d98b79fa2353b6acbf63f61190dfd1a8bde5be8d