Initial prototype of a "assistant" worker
If a worker is created using assistant=True, then it will pull *any* tasks from the scheduler,
not just tasks that are assigned to it.

The biggest limitation atm is that the assistant needs to be aware of what those tasks map to,
i.e. somehow they have to be imported for the worker to know what to do with the tasks.

It also doesn't handle failure very well.


Former-commit-id: e6c7635a071396e45a64a16d2de4ea62d8bf21bc