Adds sudo user to worker info when available
Many of the jobs I use have to be run as hadoop, so it's impossible to tell from
the user who actually ran the job. This adds sudo_user to the worker info so
that it's easier to tell who's running what. This is very handy when I need to
figure out why a large backfill or test job is running if it's causing a lot of
errors or hogging all the resources.


Former-commit-id: 5339987a59ce1a9b18a832197d62df74ded8fcbe