Updates documentation for configuration
This updates the documentation to contain explanations of most configuration
options that can be specified in client.cfg. I tried to include as many as
possible, but I've probably missed some. I grepped the code to try to find all
of them. For completeness, these include my best understanding of how each
parameter works. Many descriptions are therefore incomplete and need to be
filled out better by people who understand them.

Some descriptions may be wrong, but only in ways that reflect the understanding
an average user would have after reading the code to figure out how to properly
configure luigi. So this shouldn't be any worse than what we already have even
where it's inaccurate.


Former-commit-id: 95bb31325658b155d3b5052a386435d1bbb04dfa