Make streaming map-reduce use atomic move_dir
This is essentially using the code introduced in spotify/luigi#393,
solving the exact same issue described there.

Without this patch, the renaming is of LBYL nature, and you could end up
with moving a directory inside another instead of actually renaming a
directory.

So others don't fall into the same pitfall, I added a comment to the
rename and move function discouraging it to be used for directories.


Former-commit-id: 3a5e8d5824adc58cfdeb32d27a246f875601326f