Limit number of shown tasks in visualiser
I've noticed that my visualiser has become unusable recently when it has too
many done tasks scheduled. The tab freezes up for a long time while processing
all the tasks, and then is very sluggish afterward. When I have this many done
tasks, it's too much to really browse through anyway, so this limits the
information shown to just the count. This trades off available information with
usability of the tool.

The maximum number of shown tasks is controlled by the scheduler and can be
adjusted in client.cfg. It may be better to make this part of the query so that
it could potentially be controlled by the end user. We may also in the future
want to add additional server queries to update larger sections for searches,
as this removes useful search information.


Former-commit-id: bd1a16837d7e87203915f8aabb5683ccb5539eba