Don't have contrib-dependencies
Currently, we can't build the luigi debian package. The reason is that
the requirements in setup.py is including dependencies that are for
packages that (at least should) belong to contrib.

I also removed requests as I didn't see it being used anywhere. The test
suite should reveal if it was required or not. :)


Former-commit-id: 5fb3e577a89902892f04b02567c0ffb9832f2229