Sets task counts properly when not visible
This fixes a bug in which loading a dependency graph or visualiser page would
lead to all counts being zero. The count updates were being unnecessarily
limited to only visible groups, so this lifts that restriction. Count update
correctness is maintained because the lines computing the counts are not
affected.


Former-commit-id: e798d8ad73d4a9eda013e25d924804fd913a0f04