Allows searching within too large task groups in the visualiser.
The visualiser does not display task groups larger than a configurable limit in
order to preserve functionality when the data is too large for the browser. This
can be very frustrating when you need to know about tasks that are in one of the
groups that is not displayed. In order to preserve access to this information,
search queries are now sent to the server so that if the query filters down a
group to a more manageable size, the results can be shown.

In order to decrease latency, server requests are only made for groups that were
originally too large to be fully displayed, and are not updated if the last
server query used a prefix of the original query and returned a result list. In
this case, we can just filter the existing results on the client side.


Former-commit-id: 4b49295c4a46ee2d9a5d9f9202a0fac845521e6b