Don't make SnakebiteClient fallback to HdfsClient
This feature is already implemented and have been running in production
for years and is tested. See get_autoconfig_client.

The feature removed in this commit is a bit weird and untested behavior.
It also only works on import errors. Which will make your code behave
differntly when run from different environments.


Former-commit-id: 0aae4901635eea7cae21123d2690161874c45933