Remove all usages of is_global
Unfortunately, I found no sensible way to keep backwards compatibility
with the `--pool` parameter. It's really unclear why it was global in
the first place, because in many many cases your --task is a subclass of
BaseHadoopJobTask anyway.

I know for a fact that the is_global was (mis)used at Spotify in the
lack of a `positional=False` option.


Former-commit-id: 5d68fd98226949bee0a5610782f413c255faef5b