Fix get_work() slowdown introduced in #986
As daveFNbuck said in that PR. The patch in #986 slowed down his
get_work() calls. My theory is that I actually worsened the worst case
complexity in that patch because I assumed that tasks-per-worker is less
than total-pending-tasks. While that's definetly is the case at Spotify.
That is not the case at Houzz.

This patch tries to get the fastest speed in both settings. The setting
that will get a speed-up from this patch is when you have a very few
workers and most of your tasks are in the DONE state.


Former-commit-id: 7ee261cf8697bfd087b5144178b3c0634337e8d8