Larger write in test_put_file
Previous `gcs_test` just wrote 'hi' to the GCS bucket file. This missed the bug in line 144 because it uploads in a single chunk. By increasing the write to a couple of chunks, the test only passes with my uploader fix in `gcs.py`. (The existing file gets a TypeError exception.)

Former-commit-id: ff39f11226ec8b95de31f0b24f4d2a1aa4746f83