Remove is_list support for Parameters
The main motivation is that they are harder to reason about. For example
it's unclear how command-line parameters and config interplay (replacing
semantics or additive semantics? If additive, in what order?)

It was proposed in this Google Group thread[1]. Indeed, this does break
backward compatibility, but it's quite a burden to maintain in addition
to having unclear semantics.

As a workaround, you can for example have comma-seperated parameters.
One can potentially implement a listify-factory for Parmeters objects
too.

[1]: https://groups.google.com/forum/#!topic/luigi-user/kywHGzDD02c


Former-commit-id: 3d70f8eb50fcead9946d07c8dbe66cb242372d28