Hide buggy parts of Parameter object API
The current implementation of Parameters is flawed and there are some
bugs. The main issue is that the Parameter objects contains state and
the same object are sometimes reused (for example by inheritance)
between task classes.

This patch doesn't solve this, but at least tries to discourage the bad
usages of parameters people might start relying on.

Users who did `MyTask.my_parameter.value` should switch over to
something like `MyTask().value`. As it's only during the instantiation
of a task that you know the task family and can match config and
parameters names (for example `--MyTask-my-parameter 123`).


Former-commit-id: d8e8c5a20f43f3c54ef8522566ab6312d1ce9e74