Deprecate SUSPENDED status
The reason is just to simplify the code, it's confusing when there are
so many statuses. And it's not obvious at all how it's utilized and that
it's only associated with the new_deps parameter.

This makes so new clients will not send the status SUSPENDED. Server
will still just convert it to PENDING as previously.

Hence, it will work for "old client new server", but not vice verse.


Former-commit-id: 188ebceeabf3e07bf5c4e9e4cd1af453bcf98916