Adds a disable worker button to the visualiser
This allows workers to be stopped from the visualiser. Workers disabled in this
way will not be able to schedule new jobs or get_work, but they are able to
finish the jobs they're currently working on. This allows for clean transitions
to newly deployed code or safe shutdown of workers when things are going wrong.

A popup verifies the intent to disable the worker to help prevent accidental
disables.


Former-commit-id: e5856880c3105f410d41fbb67f52a63a3f87914e