Shortens task graphs by counting each node only once
In computeDepth in graph.js, the entire dependency tree is traversed and each
node is given a depth and position within its depth. This computation is a bit
off, however, as a node may be encountered multiple times during traversal. This
could be easily fixed by stopping the recursion when we hit a seen node, but
then we could end up with a node depending on a node at a higher level than
itself.

Instead, I turn this into a two-stage affair. In the first stage, we do the full
traversal and only remember the maximum depth for each node. In the second stage
we traverse the whole tree again and only place nodes when we hit them for the
first time at the right depth. The second step requires a tree traversal rather
than just iterating over the nodes array because it groups child nodes together
and results in a much more readable and nicer-looking graph.

This shrinks each level's height by a factor of about the average in-degree.

Former-commit-id: 4bf7b2210432e7406fd4fbb18a531565bee1bf47