Shows spinner on status info boxes while loading
It can be a little confusing to tell whether the info boxes on the visualiser
tasks tab have been updated yet or not when using Filter on Server in the
datatables search box. This makes it more clear by resetting the number to a
question mark on updates and also turning the icon into a spinner while it's
waiting to load.


Former-commit-id: 1bf7c8bb32adeda08d0c81cfa075af7545b0fe9b