Allows limiting of graph size returned by scheduler
Like with task lists, excessively large graphs can render the visualiser
unusable. To prevent this issue, we've added a max-graph-nodes parameter that
cuts off the graph after enough nodes are seen.

In order to achieve this, the graph recursion had to be re-written as a an
iterative breadth-first traversal. The inverse dependency graph function was
re-written to be able to call this same traversal, resulting in greatly improved
complexity. Rather than scanning the entire graph once per node to find its
reverse neighbors, we scan the whole graph exactly once and store the reverse
edges in a lookup table. This improves the complexity from number of nodes times
size of graph to just number of nodes.


Former-commit-id: 4dcdc38f2478b52b5b57cd50ae8e4be5680d03cf