AWS EC2 Container Service support
From the AWS website,

> Amazon EC2 Container Service (ECS) is a highly scalable, high
performance
> container management service that supports Docker containers and
allows you
> to easily run applications on a managed cluster of Amazon EC2
instances.

To use ECS, you create a "task definition" JSON that defines the
`docker run`
command for one or more containers in a task or service, and then
submit this
JSON to the API to run the task.

This wrapper allows you to create Luigi Tasks to submit ECS
`taskDefinition`s.
You can either pass a dict (mapping directly to the `taskDefinition`
JSON) or an
Amazon Resource Name (arn) for a previously registered `taskDefinition`.

Testing was performed on the default cluster created through the ECS
console
[first run wizard](https://console.aws.amazon.com/ecs/home#/firstRun)

Requires boto3, configured with valid AWS credentials


Former-commit-id: a9bca5fa05433e94b94592afef646b67cb2a929b