Overrides scheduler disable with manual disable by worker
When the scheduler disables a task, we then ignore most other status changes
requested by the worker until re-enable. This had a weird result for me recently
when a group of tasks were repeatedly failing and many of them were disabled.
In order to stop the repeated runs, I marked them all as disabled. A few hours
later when the disable period expired for the automatic scheduler disables,
many of the tasks became PENDING again and started running. It took a while
before enough of them were scheduled with the right timing to become permanently
disabled.

To fix this, we now allow the permanent disable requested by the worker to
override the temporary disable triggered within the scheduler. Note that this
was the original behavior but it wasn't enforced in unit tests and got lost in a
refactor.

Former-commit-id: c521d59c5eacf6c19ce3c17a62f73e042fa0556e