Range helpers: Honor param_name when inferring fs
Range* allows you to specify the name of the parameter that is the date
the range is about. Another orthogonal feature is that you don't need to
provide a bulk_complete for the most common file system use case.

However, when using both these features doesn't work. This patch fixes
that.

While changing this code. I also realized that we're redundantly passing
around the task_cls even though `self` already contains it. We change
this and add a backward compatible wrapper that spews out a deprecation
warning to nudge users to update their code.


Former-commit-id: c0987f7d76beebd5d6c2e4e4c991587e86bea6eb