Improves display of unknown tasks
Rather than showing UNKNOWN as the family for an unknown task id in the
visualiser graph, we can usually successfully infer the family from the task id.
This is far preferable to just showing UNKNOWN, as it provides useful
information while looking at dependency graphs. Including the full task id in
the params dict allows for a little more information to be gleaned.


Former-commit-id: 09822353dcce9797039a3c6cbe0c80703df904fd