Implement pass-through parameters on Range tasks (#1675)
Parameters to tasks wrapped by RangeBase and its subclasses can now be
passed through the Range to the 'of' task. This both makes it easier to
specify parameters to tasks running inside of ranges in Python code and
allows for Luigi to differentiate between multiple Range instances
running with the same 'of' parameter.

Modified bulk_complete invocations to pass of_params

In RangeDaily and RangeHourly, the bulk_complete parameter was invoked
with a list of the date/hour parameters to be used for the of task. This
prevents of_params from being provided to the bulk_complete code.
Instead, we change the invocation of the bulk_complete method to use a
partial with the of_params values curried. Thus, any code using the cls
parameter of bulk_complete can behave as normal while still allowing the
of_params to be passed through.

For code that would rather subclass and override missing_datetimes, two
new helper methods are provided: datetime_to_parameters and
parameters_to_datetime. These are parallel to datetime_to_parameter and
parameter_to_datetime and rather than handling only the first argument
to the of class, handle all the required arguments.

Former-commit-id: 859095a106864ef11176fbe03c3bb8f8a4c3ad42